refactor to typescript
