--- _config.yml
@@ -1,8 +1,0 @@
-title: React-avatar-editor
-description: 
-google_analytics: 
-show_downloads: true
-theme: jekyll-theme-cayman
-
-gems:
-  - jekyll-mentions
