--- bower.json
@@ -1,20 +1,0 @@
-{
-  "name": "react-avatar-editor",
-  "description": "Facebook like avatar / profile picture component. Resize and crop your uploaded image using a intuitive user interface.",
-  "main": "dist/index.js",
-  "authors": [
-    "Moritz \"mosch\" Schwoerer"
-  ],
-  "license": "MIT",
-  "keywords": [
-    "react-component",
-    "reactjs",
-    "react",
-    "canvas",
-    "avatar",
-    "editor",
-    "profile",
-    "image"
-  ],
-  "homepage": "https://github.com/mosch/react-avatar-editor"
-}
