+++ example/index.html
@@ -1,5 +1,6 @@
 <!doctype html>
 <html>
+
 <head>
   <title>React-avatar-editor</title>
 
@@ -14,6 +15,7 @@
 
   <!--[if lt IE 9]><script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script><![endif]-->
 </head>
+
 <body>
   <div class="wrapper">
     <header>
@@ -43,7 +45,7 @@
       </h3>
 
       <h2>Demo time!</h2>
-      <div id="app"></div>
+      <div id="root"></div>
 
       <h3>
         <a id="support-or-contact" class="anchor" href="#support-or-contact" aria-hidden="true">
@@ -63,6 +65,7 @@
     </footer>
   </div>
   <!--[if !IE]><script>fixScale(document);</script><![endif]-->
-  <script src="bundle.js"></script>
+  <script src="./index.tsx"></script>
 </body>
+
 </html>
