+++ src/utils/load-image-url.ts
@@ -1,0 +1,24 @@
+/* eslint-env browser, node */
+
+function isDataURL(url: string) {
+  if (url === null) {
+    return false;
+  }
+  const regex = /^\s*data:([a-z]+\/[a-z]+(;[a-z-]+=[a-z-]+)?)?(;base64)?,[a-z0-9!$&',()*+;=\-._~:@/?%\s]*\s*$/i;
+  return !!url.match(regex);
+}
+
+export default function loadImageURL(
+  imageURL: string,
+  crossOrigin = ''
+): Promise<HTMLImageElement> {
+  return new Promise((resolve, reject) => {
+    const image = new Image();
+    image.onload = () => resolve(image);
+    image.onerror = reject;
+    if (isDataURL(imageURL) === false && crossOrigin) {
+      image.crossOrigin = crossOrigin;
+    }
+    image.src = imageURL;
+  });
+}
