+++ src/utils/retrieve-image-url.test.ts
