+++ src/utils/retrieve-image-url.ts
