+++ src/utils/to-radians.ts
@@ -1,0 +1,3 @@
+// helpers to calculate vectors
+const toRadians = (degree: number) => degree * (Math.PI / 180);
+export default toRadians;
